+++ src/components/display/view-switcher/ViewSwitcherToggle.tsx
@@ -13,9 +13,9 @@ import {Toggle, ListProps} from "../../ui"
 
 const defaults = require("lodash/defaults")
 
-
 export interface ViewSwitcherProps extends SearchkitComponentProps {
 	listComponent?:RenderComponentType<ListProps>
+  switchView?:any
 }
 
 export class ViewSwitcherToggle extends SearchkitComponent<ViewSwitcherProps, any> {
@@ -34,6 +34,7 @@ export class ViewSwitcherToggle extends SearchkitComponent<ViewSwitcherProps, an
 
   setView(view){
     this.getViewOptionsSwitcherAccessor().setView(view)
+    this.props.switchView(view)
   }
 
   render() {
@@ -41,6 +42,9 @@ export class ViewSwitcherToggle extends SearchkitComponent<ViewSwitcherProps, an
     if(viewOptionsAccessor){
       let options = viewOptionsAccessor.options
       let selectedOption = viewOptionsAccessor.getSelectedOption().key
+
+      this.props.switchView(selectedOption)
+
 			return renderComponent(this.props.listComponent, {
 				mod:this.props.mod,
 				className:this.props.className,