+++ src/components/display/view-switcher/ViewSwitcherToggleContainer.ts
@@ -1,0 +1,14 @@
+const connect = require('react-redux').connect
+import { ViewSwitcherToggle } from './ViewSwitcherToggle'
+import { switchView } from './actions'
+
+const mapDispatchToProps = dispatch => ({
+  switchView(view) {
+    dispatch(switchView(view))
+  }
+})
+
+export const ViewSwitcherToggleContainer = connect(
+  undefined,
+  mapDispatchToProps
+)(ViewSwitcherToggle)