+++ src/components/index.ts
@@ -5,6 +5,7 @@ export * from "./search/filters/selected-filters";
 export * from "./search/filters/hierarchical-menu-filter/src/HierarchicalMenuFilter";
 export * from "./search/filters/numeric-refinement-list-filter/src/NumericRefinementListFilter";
 export * from "./search/hits/src/Hits";
+export * from "./search/hits/src/HitsContainer";
 export * from "./search/hits/src/NoHits";
 export * from "./search/hits/src/InitialLoader";
 export * from "./search/hits-stats/src/HitsStats";