+++ src/core/accessors/HighlightAccessor.ts
@@ -1,26 +1,14 @@
 import {Accessor} from "./Accessor";
-const mapValues = require("lodash/mapValues")
-const zipObject = require("lodash/zipObject")
-const constant = require("lodash/constant")
-
 
 export class HighlightAccessor extends Accessor {
 
-  highlightFields:any
-  constructor(public fields:Array<string>){
+  highlightOptions:any
+  constructor(public options:Object){
     super()
-    this.highlightFields = this.computeHighlightedFields(fields)
+    this.highlightOptions = options
   }
 
-  computeHighlightedFields(fields) {
-    return {
-      fields:mapValues(
-        zipObject(fields),
-        constant({})
-      )
-    }
-  }
   buildOwnQuery(query){
-    return query.setHighlight(this.highlightFields)
+    return query.setHighlight(this.highlightOptions)
   }
 }