+++ webpack.config.js
@@ -28,22 +28,22 @@ module.exports = {
     new webpack.optimize.DedupePlugin(),
     new webpack.optimize.OccurenceOrderPlugin(),
     new ExtractTextPlugin("theme.css", {allChunks:true}),
-    new webpack.optimize.UglifyJsPlugin({
-      mangle: {
-        except: ['require', 'export', '$super']
-      },
-      compress: {
-        warnings: false,
-        sequences: true,
-        dead_code: true,
-        conditionals: true,
-        booleans: true,
-        unused: true,
-        if_return: true,
-        join_vars: true,
-        drop_console: true
-      }
-    })
+    // new webpack.optimize.UglifyJsPlugin({
+    //   mangle: {
+    //     except: ['require', 'export', '$super']
+    //   },
+    //   compress: {
+    //     warnings: false,
+    //     sequences: true,
+    //     dead_code: true,
+    //     conditionals: true,
+    //     booleans: true,
+    //     unused: true,
+    //     if_return: true,
+    //     join_vars: true,
+    //     drop_console: true
+    //   }
+    // })
   ],
   externals: {
     "react": "React",